@@ -1,7 +1,7 @@
 
 module.exports = {
   content: [
-    './app/**/*.{js,ts,jsx,tsx,mdx}',
+    './src/app/**/*.{js,ts,jsx,tsx,mdx}',
     './src/components/**/*.{js,ts,jsx,tsx,mdx}',
   ],
   plugins: [require('@tailwindcss/forms'), require('@tailwindcss/typography')],
